Added QPRK solvers
