+++ lib/OrdinaryDiffEqQPRK/src/alg_utils.jl
@@ -1,0 +1,1 @@
+alg_order(alg::QPRK98) = 9
